Made the most minimal version of the non-working l4casadi example
